requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
python-dotenv==1.0.1

# Phase 2 - Document Parsing
pdfminer.six==20231228
python-docx==1.1.2
pymupdf==1.24.0

# Phase 3 - LLM
anthropic==0.28.0

# Phase 4 - WCAG / Output
openpyxl==3.1.2

# AI model
paddlepaddle==3.3.1
paddlex==3.4.3
